+++ apps/web/src/components/DataTable/DataTable.tsx
@@ -11,36 +11,6 @@ type DataTableProps = {
   data: DataRow[];
 };
 
-function getSortableValue(value: string) {
-  const trimmedValue = value.trim();
-
-  if (!trimmedValue) {
-    return trimmedValue;
-  }
-
-  const numericValue = Number(trimmedValue);
-
-  if (!Number.isNaN(numericValue) && Number.isFinite(numericValue)) {
-    return numericValue;
-  }
-
-  return trimmedValue.toLowerCase();
-}
-
-function compareValues(left: string, right: string) {
-  const leftValue = getSortableValue(left);
-  const rightValue = getSortableValue(right);
-
-  if (typeof leftValue === 'number' && typeof rightValue === 'number') {
-    return leftValue - rightValue;
-  }
-
-  return String(leftValue).localeCompare(String(rightValue), undefined, {
-    numeric: true,
-    sensitivity: 'base',
-  });
-}
-
 function DataTable({ columns, data }: Readonly<DataTableProps>) {
   const searchFieldId = useId();
   const [searchTerm, setSearchTerm] = useState('');
@@ -48,59 +18,9 @@ function DataTable({ columns, data }: Readonly<DataTableProps>) {
   const [sortDirection, setSortDirection] = useState<SortDirection>('asc');
   const isTableReset = !searchTerm.trim() && sortKey === null;
 
-  if (!columns.length || !data.length) {
-    return (
-      <TableState
-        variant="empty"
-        title="No table data"
-        message="There is no data to show in this table yet."
-      />
-    );
-  }
+  
 
-  const normalizedSearchTerm = searchTerm.trim().toLowerCase();
-
-  const filteredRows = data.filter((row) => {
-    if (!normalizedSearchTerm) {
-      return true;
-    }
-
-    return columns.some((column) =>
-      (row[column.key] ?? '').toLowerCase().includes(normalizedSearchTerm)
-    );
-  });
-
-  const sortedRows = [...filteredRows];
-
-  if (sortKey) {
-    sortedRows.sort((leftRow, rightRow) => {
-      const comparison = compareValues(leftRow[sortKey] ?? '', rightRow[sortKey] ?? '');
-
-      return sortDirection === 'asc' ? comparison : comparison * -1;
-    });
-  }
-
-  const handleSort = (columnKey: string) => {
-    if (sortKey === columnKey) {
-      setSortDirection((currentDirection) =>
-        currentDirection === 'asc' ? 'desc' : 'asc'
-      );
-      return;
-    }
-
-    setSortKey(columnKey);
-    setSortDirection('asc');
-  };
-
-  const handleResetTable = () => {
-    setSearchTerm('');
-    setSortKey(null);
-    setSortDirection('asc');
-  };
-
-  const resultCountLabel = `${sortedRows.length} ${
-    sortedRows.length === 1 ? 'row' : 'rows'
-  } shown`;
+  
 
   return (
     <section className={styles.wrapper}>
